Add option to alert on disabled task, perfdata fix
This commit adds functionality to alert when any matching task is marked
as disabled when the flag --AlertOnDisabled is passed to the script.

It also fixes the format of the perfdata to better fit the specification
as per https://www.monitoring-plugins.org/doc/guidelines.html#AEN201,
allowing e.g. Graphios to parse it. Additionally, 'Disabled Tasks' is
added to the perfdata output.